+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+
+## [1.4.2](https://github.com/kb-dk/ds-license/releases/tag/ds-license-1.4.2) 2024-06-11
 - Added YAML property to disable all GUI: gui.adminGuiEnabled=true
 - Added YAML property for cache refresh time : cache.reloadInSeconds=300
 - Audit log will now show name and email from AD loaded by KeyCloak.